requirements.txt
sqlalchemy>=2.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
flake8>=6.0.0
pytest-cov>=4.0.0